// src/app/modules/setup/models/item-profile-categories-price.ts
export class ItemProfileCategoriesPrice {
    ID: number = 0;
    ItemID: number = null;
    ItemCategoryID: number = null;
    SalePcs: number = null;
    SalePack: number = null;
    SaleCtn: number = null;
}

export class VW_ItemProfileCategoriesPrice {
    ID: number = 0;
    ItemID: number = null;
    CategoryDescription: number = null;
    ItemCategoryID: number = null;
    SalePcs: number = null;
    SalePack: number = null;
    SaleCtn: number = null;
    IsSaleAmount: boolean = false;
}